requirements.txt
Flask==3.1.0
google-generativeai==0.8.4
pydub==0.25.1
SpeechRecognition==3.14.2
gunicorn